Flatten package tree
The game and player package had to be merged, the reason is as
following:

TL;DR: we need references to players in game objects and references to
games in player objects.

The game objects needs to have some sort of collection of players, to
have a collection of players we need to have access to the player id
type or the player type(to store a pointer to an instance of it).

The player objects needs to have a reference to its game, to have a
reference to the game we need access to either the game id type or the
game type(to store a pointer to an instance of it).

player and game has to live in the same package since otherwise we
would have a import/dependency cycle.